refactor: PaymentCard 컴포넌트 util 패키지로 이동
